refactor: unify EvalsReport and RunResult types
